+++ app/src/API/Interfaces/SqlExecutor/Forked/ShellExecutor.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Db3v4l\API\Interfaces\SqlExecutor\Forked;
+
+use Symfony\Component\Process\Process;
+
+interface ShellExecutor
+{
+    /**
+     * @return Process
+     */
+    public function getExecuteShellProcess();
+}